Quote URL has changed since implementation and stocks can be non integer values

+++ quote.go
@@ -2,9 +2,9 @@ package finnhub
 
 // Quote the data structure for quotes
 type Quote struct {
-	Open          int `json:"o"`
-	High          int `json:"h"`
-	Low           int `json:"l"`
-	Current       int `json:"c"`
-	PreviousClose int `json:"pc"`
+	Open          float64 `json:"o"`
+	High          float64 `json:"h"`
+	Low           float64 `json:"l"`
+	Current       float64 `json:"c"`
+	PreviousClose float64 `json:"pc"`
 }
